+++ src/components/checkin/index.ts
@@ -5,3 +5,5 @@
 
 export { default as ObstacleBottomSheet } from './ObstacleBottomSheet';
 export { default as SuccessAnimation } from './SuccessAnimation';
+export { default as EditHabitBottomSheet } from './EditHabitBottomSheet';
+export type { HabitWordingUpdates } from './EditHabitBottomSheet';